Rename old to deprecated sort-by-year
