+++ frontend/app/components/activities-edit.js
@@ -1,5 +1,5 @@
 import Component from "@ember/component";
-import sortByYear from "../utils/sort-by-year-old";
+import sortByYear from "../utils/deprecated-sort-by-year";
 import { inject as service } from "@ember/service";
 import { on } from "@ember/object/evented";
 import { EKMixin, keyUp } from "ember-keyboard";
